+++ CHANGELOG.md
@@ -62,6 +62,7 @@ end
 - **decidim-proposals**: Update card layout [\#3338](https://github.com/decidim/decidim/pull/3338)
 - **decidim-debates**: Update card layout [\#3371](https://github.com/decidim/decidim/pull/3371)
 - **decidim-participatory_processes**: Update card layout for processes [\#3382](https://github.com/decidim/decidim/pull/3382)
+- **decidim-participatory_processes**: Update card layout for process groups [\#3395](https://github.com/decidim/decidim/pull/3395)
 
 **Fixed**: